gallery tab size and produts tab size fized

+++ src/firebase/api.ts
@@ -185,6 +185,7 @@ export const createStore = async (uid: string, payload: any) => {
     await setDoc(doc(db, "store", storeId), {
       ...payload,
       userId: uid,
+      id: storeId,
       active: false,
       createdAt: new Date(),
       published: false,
@@ -193,19 +194,34 @@ export const createStore = async (uid: string, payload: any) => {
       visitCount: 0,
       verified: false,
       showProfile: false,
+      haveUpdate: [],
+      gallery: [],
+      location: "",
+      companyProfilePdfUrl: "",
+      youtubeVideos: [],
     });
 
     for (let index = 0; index < 4; index++) {
       console.log("payload");
       try {
-        await addDoc(collection(db, "store", storeId, "top-slider"), {
+        const { id } = await addDoc(
+          collection(db, "store", storeId, "top-slider"),
+          {
+            imageUrl: "",
+          }
+        );
+
+        await setDoc(doc(db, "latestStore", storeId, "top-slider", id), {
           imageUrl: "",
         });
       } catch (error) {
         console.log(error);
       }
     }
+
     console.log("Document successfully written to Firestore!");
+
+    return storeId;
   } catch (error) {
     console.error("Error writing document:", error);
   }
@@ -243,6 +259,70 @@ export const updateStore = async (storeId: string, payload: any) => {
     });
 
     console.log("Document Update successfully");
+  } catch (error) {
+    console.error("Error writing document:", error);
+  }
+};
+
+// ------------------------------------------------------
+
+export const updateStore2 = async (storeId: string, payload: any) => {
+  // console.log(payload);
+
+  try {
+    await updateDoc(doc(db, "store", storeId), {
+      ...payload,
+    });
+
+    console.log("Document Update successfully");
+  } catch (error) {
+    console.error("Error writing document:", error);
+  }
+};
+
+// ---------------------------------------------
+
+export const updateStore3 = async (storeId: string, payload: any) => {
+  // console.log(payload);
+
+  try {
+    await updateDoc(doc(db, "latestStore", storeId), {
+      ...payload,
+    });
+
+    // const collectionRef = collection(db, "store", storeId, "top-slider");
+    // const querySnapshot = await getDocs(collectionRef);
+    // const documentsExist = !querySnapshot.empty;
+
+    // if (!documentsExist) {
+    //   for (let index = 0; index < 4; index++) {
+    //     console.log("payload");
+    //     try {
+    //       await addDoc(collection(db, "latestStore", storeId, "top-slider"), {
+    //         imageUrl: "",
+    //       });
+    //     } catch (error) {
+    //       console.log(error);
+    //     }
+    //   }
+    // }
+
+    console.log("Document Update successfully");
+  } catch (error) {
+    console.error("Error writing document:", error);
+  }
+};
+
+// --------------------------------------------
+export const syncLatestStoreWithStore = async (storeId: string) => {
+  // console.log(payload);
+
+  try {
+    const documentRef = doc(db, "store", storeId);
+    const querySnapshot = await getDoc(documentRef);
+
+    const latestDocumentRef = doc(db, "latestStore", storeId);
+    await updateDoc(latestDocumentRef, {...querySnapshot.data()});
   } catch (error) {
     console.error("Error writing document:", error);
   }
@@ -499,6 +579,20 @@ export const togglePublish = async (storeId: string, published: boolean) => {
     await updateDoc(documentRef, {
       published: !published,
     });
+    // syncLatestStoreWithStore(storeId);
+     try {
+       const documentRef = doc(db, "store", storeId);
+       const querySnapshot = await getDoc(documentRef);
+
+       // const storeData = querySnapshot.data() as StoreObj
+
+       const latestDocumentRef = doc(db, "latestStore", storeId);
+       await updateDoc(latestDocumentRef, {
+         published: querySnapshot.data()?.published,
+       });
+     } catch (error) {
+       console.error("Error writing document:", error);
+     }
   } catch (error) {
     console.log(error);
   }
@@ -519,23 +613,24 @@ export const togglePublish = async (storeId: string, published: boolean) => {
 // };
 
 export const createMessageToAll = async (message: string) => {
-  const collectionRef = collection(db, "users");
-  const querySnapshot = await getDocs(collectionRef);
+  try {
+    const collectionRef = collection(db, "users");
+    const querySnapshot = await getDocs(collectionRef);
 
-  const adminMessagesCollectionRef = collection(db, "amdinMessages");
-  const { id: adminMessageId } = await addDoc(adminMessagesCollectionRef, {
-    message,
-    createdAt: new Date(),
-    imageUrl: "",
-    fromName: "admin",
-    fromId: "",
-    toName: "all",
-    toId: "",
-    seen: false,
-  });
+    const adminMessagesCollectionRef = collection(db, "adminMessages");
+    const { id: adminMessageId } = await addDoc(adminMessagesCollectionRef, {
+      message,
+      createdAt: new Date(),
+      imageUrl: "",
+      fromName: "admin",
+      fromId: "",
+      toName: "all",
+      toId: "",
+      seen: false,
+    });
 
-  querySnapshot.forEach(async (doc) => {
-    try {
+    // Create an array of promises for adding messages to each user
+    const addMessagePromises = querySnapshot.docs.map(async (doc) => {
       const userMessagesRef = collection(doc.ref, "messages");
       await addDoc(userMessagesRef, {
         message,
@@ -548,11 +643,16 @@ export const createMessageToAll = async (message: string) => {
         toId: "",
         seen: false,
       });
-    } catch (error) {
-      console.log("Error adding message to user:", doc.id, error);
-      throw error;
-    }
-  });
+    });
+
+    // Wait for all messages to be added to users
+    await Promise.all(addMessagePromises);
+
+    console.log("Message sent to all users successfully.");
+  } catch (error) {
+    console.log("Error sending message to all users:", error);
+    throw error;
+  }
 };
 
 // -------------------------------------------------------
@@ -572,12 +672,12 @@ export const createMessageToUser = async (message: string, email: string) => {
 
   if (!userIdForSetMessage) return "There are no matching user";
 
-  const userMessagesRef = collection(
-    db,
-    "users",
-    userIdForSetMessage,
-    "messages"
-  );
+  // const userMessagesRef = collection(
+  //   db,
+  //   "users",
+  //   userIdForSetMessage,
+  //   "messages"
+  // );
 
   try {
     const adminMessagesCollectionRef = collection(db, "adminMessages");
@@ -587,22 +687,27 @@ export const createMessageToUser = async (message: string, email: string) => {
       imageUrl: "",
       fromName: "admin",
       fromId: "",
-      toName: matchingUser.id,
-      toId: "",
+      toName: email,
+      toId: matchingUser.id,
       seen: false,
     });
 
-    await addDoc(userMessagesRef, {
-      message,
-      messageId: adminMessageId,
-      createdAt: new Date(),
-      imageUrl: "",
-      fromName: "admin",
-      fromId: "",
-      toName: matchingUser.id,
-      toId: userIdForSetMessage,
-      seen: false,
-    });
+    console.log(adminMessageId);
+
+    await setDoc(
+      doc(db, "users", userIdForSetMessage, "messages", adminMessageId),
+      {
+        message,
+        messageId: adminMessageId,
+        createdAt: new Date(),
+        imageUrl: "",
+        fromName: "admin",
+        fromId: "",
+        toName: matchingUser.id,
+        toId: userIdForSetMessage,
+        seen: false,
+      }
+    );
     console.log("New Message added..");
   } catch (error) {
     console.log(error);
@@ -631,9 +736,16 @@ export const updateAsSeen = async (uid: string) => {
 
 // ------------------------------------
 
-export const handleMessageDelete = async (id: string) => {
+export const handleMessageDelete = async (
+  userId: string,
+  messageId: string
+) => {
+  if (!userId && !messageId) return;
   try {
-    const documentRef = doc(db, "messagesToAll", id);
+    const documentRef1 = doc(db, "adminMessages", messageId);
+    await deleteDoc(documentRef1);
+
+    const documentRef = doc(db, "users", userId, "messages", messageId);
     await deleteDoc(documentRef);
     console.log("Message deleted successfully");
   } catch (error) {
